requirements.txt
django==3.2.13
gitpython
pymysql
#django_crontab
requests
django_apscheduler